+++ src/ext/commands.py
@@ -975,6 +975,10 @@ Ensure the bot has these permissions for smooth operation.
                      value=f"```{translated}```", inline=False)
         await interaction.followup.send(embed=em, ephemeral=True)  # noqa
 
+    @app_commands.command(name="report_incorrect_chapters")
+    async def report_incorrect_chapters(self, interaction: discord.Interaction):
+        await interaction.response.defer(ephemeral=True, thinking=True)
+
 
 async def setup(bot: MangaClient) -> None:
     if bot.debug and bot.test_guild_ids: